requirements.txt
tensorflow==2.12.0
tensorflow-hub==0.13.0
opencv-python==4.8.1.78
numpy==1.23.5
pandas==2.0.3
matplotlib==3.7.2
scikit-learn==1.3.0
nltk==3.8.1
pycocotools==2.0.7
Pillow==10.0.1
tqdm==4.66.1
PyYAML==6.0.1
fastapi==0.103.2
uvicorn[standard]==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
aiofiles==23.2.1
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.25.0
tensorboard==2.12.3
